+++ gulpfile.js/tasks/images.js
@@ -2,12 +2,19 @@
 
 const path = require('path')
 const gulp = require('gulp')
-const gutil = require('gulp-util')
 const config = require('../config')
 
-gulp.task('images', () => {
-  const dest = path.join(config.dest[gutil.env.version], config.images.destDirName)
+const images = (v) => {
+  const dest = path.join(config.snapshotDir[v], config.images.destDirName)
 
   return gulp.src(config.images.src)
     .pipe(gulp.dest(dest))
+}
+
+gulp.task('images:v3', () => {
+  return images('v3')
+})
+
+gulp.task('images:v4', () => {
+  return images('v4')
 })