+++ gulpfile.js/tasks/release.js
@@ -3,15 +3,17 @@
 const gulp = require('gulp')
 const runSequence = require('run-sequence')
 
-gulp.task('release', ['changelog'], (done) => {
+runSequence.options.showErrorStackTrace = false
+
+gulp.task('release', ['changelog', 'clean:dist'], (done) => {
   runSequence(
-    'release:v3',
-    'release:v4',
+    'build',
+    'zip',
     done
   )
 })
 
-gulp.task('release:v3', (done) => {
+gulp.task('release:v3', ['changelog', 'clean:dist'], (done) => {
   runSequence(
     'build:v3',
     'zip',
@@ -19,7 +21,7 @@ gulp.task('release:v3', (done) => {
   )
 })
 
-gulp.task('release:v4', (done) => {
+gulp.task('release:v4', ['changelog', 'clean:dist'], (done) => {
   runSequence(
     'build:v4',
     'zip',