+++ gulpfile.js/tasks/style.js
@@ -10,9 +10,8 @@ const plumber = require('gulp-plumber')
 const replace = require('gulp-replace')
 const gutil = require('gulp-util')
 const rename = require('gulp-rename')
-const runSequence = require('run-sequence')
 
-function sassToCss (stream) {
+function sassToCss (stream, version) {
   return stream
     .pipe(sourceMaps.init())
     .pipe(plumber((error) => {
@@ -23,23 +22,15 @@ function sassToCss (stream) {
     .pipe(autoprefixer({browsers: ['last 2 versions', 'IE >= 8']}))
     .pipe(rename({suffix: '.min'}))
     .pipe(sourceMaps.write(config.sass.sourceMapsDir))
-    .pipe(gulp.dest(path.join(config.dest[gutil.env.version], config.sass.destDirName)))
+    .pipe(gulp.dest(path.join(config.snapshotDir[version], config.sass.destDirName)))
 }
 
-gulp.task('style', (done) => {
-  runSequence(
-    'style:v3',
-    'style:v4',
-    done
-  )
+gulp.task('style:v3', () => {
+  const src = gulp.src(config.sass.src)
+  return sassToCss(src, 'v3')
 })
 
-gulp.task('style:v3', ['v3', 'stylelint'], () => {
-  return sassToCss(gulp.src(config.sass.src))
-})
-
-gulp.task('style:v4', ['v4', 'stylelint'], () => {
-  return sassToCss(
-    gulp.src(config.sass.src).pipe(replace(/^@import.*scss\/.*$/gmi, ''))
-  )
+gulp.task('style:v4', () => {
+  const src = gulp.src(config.sass.src).pipe(replace(/^@import.*scss\/.*$/gmi, ''))
+  return sassToCss(src, 'v4')
 })